requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
google-generativeai
protobuf<4